Added Room Functionality For The Customer --Konstantinos
I added all the necessary room functionality for the customer, which includes functionallity for the room lighting, temperature, the room's radio and the room's television. All necessary warnings for unsaved changes are also in place. Finally I changed a bit the database and dataaccess to give the room television a more extensive functionallity.