Added Initial Functionalitz For Restaurant --Konstantinos
I add the necessary functionality for the restaurant for the customers. After this update the filters are fully functional(the search could be improved) and most of the foods are present in the restaraunt at least for the evening. The foods of the restaurant for the morning view are not added yet and they will be added in a future update. The shopping card(making orders) and the food pop ups are not fully functional yet.