Fully Finished The Restaurant/Cart/Orders Forms For The Customer And Improved The Responsiveness Of The App --Konstantinos
I finished the restaurant, cart and orders form, which are pretty similar, so a lot of code was copy pasted from one to form to another, but each one was a bit different and yet I decided to thematically commit them in one update. Additionally I improved the responsiveness of the app utilizing the anchor attribute that exists in each control. Finally a bug that is extemely hard to fix has been discovered(error: "database is locked"), which occurs randomly upon the creation of an order. I have tried to resolve that bug for a day and a half and while I have restricted it, it still exists, but I have made sure that it does not crash the app and only the creation of the order which can then be submited by the user again. In a future update I will try to fix the issue.