Improved The Employee SideBar And Added Functionality For The Employee(Rooms And Pools) --Konstantinos
I changed a bit the employee sidebar and what child form it opens when the employee clicks on the rooms and the pools. As it stands the employee now has an additional choice of the pool or the room they want to change, since they have access to everything. Additionally I fixed some bugs that existed for the room interactive form for the customer and the employee. In the next update employee functionality for orders will be added.